+++ public/index.html
@@ -25,7 +25,6 @@
       Learn how to configure a non-root public URL by running `npm run build`.
     -->
     <title>React App</title>
-    
   </head>
   <body>
     <noscript>You need to enable JavaScript to run this app.</noscript>
