+++ src/index.js
@@ -2,19 +2,12 @@ import React from 'react';
 import ReactDOM from 'react-dom/client';
 import './index.css';
 import App from './App';
-import { Auth0Provider } from '@auth0/auth0-react'
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
 root.render(
-  <Auth0Provider
-  domain="dev-4nwglgqmkknvqj0p.us.auth0.com"
-  clientId="xunGi4vccXLByIuBTzUaF6R47BWeRNvd"
-  redirectUri={window.location.origin}
-  >
   <React.StrictMode>
     <App />
   </React.StrictMode>
-  </Auth0Provider>
 );
